Handlebars Additions
Added new handlebars files based on list I created for necessary views.
Put content into allMoviesPage handlebars file and am using the
bootstrap3 "card" class as my movie tile. These will be iterated using
the "each" function in handlebars. I gave the "view movie" button a
data-id of the movie.id so that I can call the correct "movie page".
I will store this data-id in the store object so that the following
view will be of the correct movie and I can use it to call all the
correct comments as well.