Handlebars
Added functionality to showMoviePage that filters movies and comments
down to an array of movies comments that only belong to to the user.
Then, edit and delete buttons use this variable in an IF statement
to iterate edit/delete buttons on movies and comments that the user
owns, and not showing them on ones the user doesn't own.

Next is making them connect to CRUD actions.
After that is form fields.